Major changes to mapping implementation, some changes to Entities. Now all mapping from entities to models and vice-versa is handles in the AutoMapperConfig file.

+++ Mhotivo/Controllers/ContactInformationController.cs
@@ -53,7 +53,7 @@ namespace Mhotivo.Controllers
         {
             var model = new ContactInformationRegisterModel
                         {
-                            Id = (int) id
+                            Id = id
                         };
             return View("ContactAdd", model);
         }
